test(import-sp): split import-engine.test.ts by phase (ps-vrac)
Split engine/import-engine.test.ts (1150 LOC, 36 tests) into 4 focused
files plus shared fixtures. Every resulting file is under 500 LOC and
the test count is preserved.

- engine/import-engine-happy-path.test.ts (421 LOC, 13 tests)
- engine/import-engine-errors.test.ts (232 LOC, 11 tests)
- engine/import-engine-resume.test.ts (106 LOC, 2 tests)
- engine/import-engine-lifecycle.test.ts (334 LOC, 10 tests)
- helpers/import-engine-fixtures.ts (93 LOC)
